+++ src/layers/ANNErrorLayer.cpp
@@ -108,7 +108,7 @@ int ANNErrorLayer::checkVertices() const {
    if (VThresh < 0 && VThresh > -0.999*max_pvvdata_t) { // 0.999 is to allow for imprecision from params files using 3.40282e+38 instead of infinity
       if (parent->columnId()==0) {
          pvErrorNoExit().printf("%s: VThresh cannot be negative (value is %f).\n",
-               getDescription_c(), VThresh);
+               getDescription_c(), (double)VThresh);
       }
       status = PV_FAILURE;
    }
